Move ProgressDialogFragment to material
