+++ collect_app/src/main/java/org/odk/collect/android/formentry/saving/FormSaveViewModel.java
@@ -26,7 +26,7 @@ import org.odk.collect.android.externaldata.ExternalDataManager;
 import org.odk.collect.android.formentry.RequiresFormController;
 import org.odk.collect.android.formentry.audit.AuditEvent;
 import org.odk.collect.android.formentry.audit.AuditUtils;
-import org.odk.collect.android.fragments.dialogs.ProgressDialogFragment;
+import org.odk.collect.material.MaterialProgressDialogFragment;
 import org.odk.collect.android.javarosawrapper.FormController;
 import org.odk.collect.android.projects.CurrentProjectProvider;
 import org.odk.collect.android.tasks.SaveFormToDisk;
@@ -56,7 +56,7 @@ import static org.odk.collect.android.tasks.SaveFormToDisk.SAVED;
 import static org.odk.collect.android.tasks.SaveFormToDisk.SAVED_AND_EXIT;
 import static org.odk.collect.shared.strings.StringUtils.isBlank;
 
-public class FormSaveViewModel extends ViewModel implements ProgressDialogFragment.OnCancelCallback, RequiresFormController, QuestionMediaManager {
+public class FormSaveViewModel extends ViewModel implements MaterialProgressDialogFragment.OnCancelCallback, RequiresFormController, QuestionMediaManager {
 
     public static final String ORIGINAL_FILES = "originalFiles";
     public static final String RECENT_FILES = "recentFiles";
