+++ collect_app/src/main/java/org/odk/collect/android/preferences/dialogs/ResetProgressDialog.kt
@@ -2,10 +2,10 @@ package org.odk.collect.android.preferences.dialogs
 
 import android.content.Context
 import org.odk.collect.android.R
-import org.odk.collect.android.fragments.dialogs.ProgressDialogFragment
+import org.odk.collect.material.MaterialProgressDialogFragment
 import org.odk.collect.strings.localization.getLocalizedString
 
-class ResetProgressDialog : ProgressDialogFragment() {
+class ResetProgressDialog : MaterialProgressDialogFragment() {
     override fun onAttach(context: Context) {
         super.onAttach(context)
