+++ material/src/main/res/layout/progress_dialog.xml
@@ -18,7 +18,7 @@
 
     <TextView
         android:layout_marginStart="@dimen/margin_standard"
-        android:textAppearance="@style/TextAppearance.Collect.Body1"
+        android:textAppearance="?textAppearanceBody1"
         android:id="@+id/message"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
